refactor media
