+++ apps/nextjs-website/src/__tests__/helpers/productHeader.helpers.test.ts
@@ -6,6 +6,8 @@ const product: Product = {
   shortName: 'test',
   description: 'test',
   logo: {
+    alternativeText: undefined,
+    caption: undefined,
     name: 'test',
     width: 60,
     height: 61,
