+++ apps/nextjs-website/src/_contents/pdnd/pdnd.ts
@@ -12,6 +12,8 @@ export const pdnd: Product = {
   slug: 'pdnd-interoperabilita',
   path: '/pdnd-interoperabilita',
   logo: {
+    alternativeText: 'PDND Interoperabilità',
+    caption: undefined,
     name: '',
     width: 60,
     height: 61,
