+++ apps/nextjs-website/src/lib/strapi/codecs/MediaCodec.ts
@@ -5,8 +5,8 @@ export const MediaAttributesCodec = t.strict({
   name: t.string,
   alternativeText: t.union([NullToUndefinedCodec, t.string]),
   caption: t.union([NullToUndefinedCodec, t.string]),
-  width: t.number,
-  height: t.number,
+  width: t.union([NullToUndefinedCodec, t.number]),
+  height: t.union([NullToUndefinedCodec, t.number]),
   ext: t.string,
   mime: t.string,
   url: t.string,
