+++ apps/nextjs-website/src/lib/types/speaker.ts
@@ -1,5 +1,5 @@
 import { BlocksContent } from '@strapi/blocks-react-renderer';
-import { Media } from './media';
+import { Media } from '@/lib/strapi/codecs/MediaCodec';
 
 export type Speaker = {
   readonly name: string;
